Applied rsc structure changes to webpack sample
Copied over from parcel sample
