+++ NavigationReact/sample/rsc-webpack/src/index.js
@@ -1,15 +1,30 @@
 import * as React from 'react';
-import { useState } from 'react';
+import { useState, useMemo } from 'react';
 import ReactDOM from 'react-dom/client';
-import { createFromFetch } from 'react-server-dom-webpack/client';
-import HmrContext from './HmrContext.js';
+import { encodeReply, createFromFetch } from 'react-server-dom-webpack/client';
+import { BundlerContext } from 'navigation-react';
+
+const fetchRSC = async (url, options) => {
+  const response = fetch(url, {
+    ...options,
+    headers: {
+      Accept: 'text/x-component',
+      ...options?.headers,
+    },
+    body: options && 'body' in options 
+      ? await encodeReply(options.body)
+      : undefined,
+  });
+  return createFromFetch(response);
+}
 
 function Shell({data}) {
   const [root, setRoot] = useState(data);
+  const bundler = useMemo(() => ({setRoot, deserialize: fetchRSC}), []);
   return (
-    <HmrContext.Provider value={setRoot}>
-        {root}
-    </HmrContext.Provider>
+    <BundlerContext.Provider value={bundler}>
+      {root}
+    </BundlerContext.Provider>
   );
 }
